api/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
joblib
scikit-learn
sqlalchemy
pymysql
python-dotenv